+++ src/services/Styles.ts
@@ -50,11 +50,24 @@ export class Styles {
     private _updateStyleSheet(): void {
         this._unloadStyleSheet();
         const themeContext = St.ThemeContext.get_for_stage(global.stage);
-        const content = this._generateStyleSheetContent();
+        let styles = this._generateStyleSheetContent();
+        this._settings.applicationStyles.value = styles;
+        if (this._settings.customStylesEnabled.value) {
+            this._settings.customStylesFailed.value = false;
+            styles = styles + '\n' + this._settings.customStyles.value;
+        }
         const [file, stream] = Gio.File.new_tmp(null);
         const outputStream = Gio.DataOutputStream.new(stream.outputStream);
-        outputStream.put_string(content, null);
-        themeContext.get_theme().load_stylesheet(file);
+        outputStream.put_string(styles, null);
+        try {
+            themeContext.get_theme().load_stylesheet(file);
+        } catch (e) {
+            console.error('Failed to load stylesheet');
+            if (this._settings.customStylesEnabled.value) {
+                this._settings.customStylesEnabled.value = false;
+                this._settings.customStylesFailed.value = true;
+            }
+        }
         outputStream.close(null);
         stream.close(null);
         file.delete(null);
@@ -146,6 +159,16 @@ export class Styles {
                 this._workspaceUpdateNotifier.notify();
             }),
         );
+        this._settings.customStylesEnabled.subscribe(() => {
+            this._updateStyleSheet();
+            this._workspacesBarUpdateNotifier.notify();
+        });
+        this._settings.customStyles.subscribe(() => {
+            if (this._settings.customStylesEnabled.value) {
+                this._updateStyleSheet();
+                this._workspacesBarUpdateNotifier.notify();
+            }
+        });
     }
 
     /** Updated style the workspaces-bar panel button. */